Fixed crash when deleting a Julia.ar object allocated before any include
This should be pushed to master aswell. When a Julia.ar object was created before an include, and then deleted after the include was performed, it was pointing to garbage memory (args and object id dict weren't allocated yet)

Former-commit-id: 9ca24c8eb54ce0b2eefa432453f007efca7a4ad2 [formerly a85b0dc2881b8e04ae0df46b2dc316022fd095c5]
Former-commit-id: 0ff895b7ea0f3feb5c07f6add10b5fff9ace3a47

+++ Julia.cpp
@@ -47,8 +47,11 @@ public:
 
     ~Julia() 
     {
-        delete_object_id_dict(global_id_dict, object_id_dict, delete_index);
-        RTFree(mWorld, args);
+        if(execute == true)
+        {
+            delete_object_id_dict(global_id_dict, object_id_dict, delete_index);
+            RTFree(mWorld, args);
+        }
     }
 
 private:
